more faffing cant make up my mind how to do this
