+++ src/Ocelot/Configuration/File/FileLoadBalancerOptions.cs
@@ -1,0 +1,9 @@
+namespace Ocelot.Configuration.File
+{
+    public class FileLoadBalancerOptions
+    {
+        public string Type { get; set; }
+        public string Key { get; set; }
+        public int ExpiryInMs { get; set; } 
+    }
+}
