+++ src/Ocelot/Configuration/LoadBalancerOptions.cs
@@ -1,0 +1,18 @@
+namespace Ocelot.Configuration
+{
+    public class LoadBalancerOptions
+    {
+        public LoadBalancerOptions(string type, string key, int expiryInMs)
+        {
+            Type = type;
+            Key = key;
+            ExpiryInMs = expiryInMs;
+        }
+
+        public string Type { get; }
+
+        public string Key { get; }
+        
+        public int ExpiryInMs { get; } 
+    }
+}
