+++ src/Ocelot/LoadBalancer/LoadBalancers/CookieStickySessions.cs
@@ -1,0 +1,87 @@
+using System;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Ocelot.Middleware;
+using Ocelot.Responses;
+using Ocelot.Values;
+
+namespace Ocelot.LoadBalancer.LoadBalancers
+{
+    public class CookieStickySessions : ILoadBalancer
+    {
+        private readonly ILoadBalancer _loadBalancer;
+        private ConcurrentDictionary<string, StickySession> _stored;
+        private Timer _timer;
+        private bool _expiring;
+        private string _key;
+        private readonly int _expiryInMs;
+
+        public CookieStickySessions(ILoadBalancer loadBalancer, string key, int expiryInMs)
+        {
+            _key = key;
+            _expiryInMs = expiryInMs;
+            _loadBalancer = loadBalancer;
+            _stored = new ConcurrentDictionary<string, StickySession>();
+            _timer = new Timer(x =>
+            {
+                if(_expiring)
+                {
+                    return;
+                }
+
+                _expiring = true;
+
+                Expire();
+                
+                _expiring = false;
+
+            }, null, 0, 50);
+        }
+
+        public async Task<Response<ServiceHostAndPort>> Lease(DownstreamContext context)
+        {
+            var value = context.HttpContext.Request.Cookies[_key];
+
+            if (!string.IsNullOrEmpty(value) && _stored.ContainsKey(value))
+            {
+                var cached = _stored[value];
+                var updated = new StickySession(cached.HostAndPort, DateTime.UtcNow.AddMilliseconds(_expiryInMs));
+                _stored[value] = updated;
+                return new OkResponse<ServiceHostAndPort>(updated.HostAndPort);
+            }
+
+            var next = await _loadBalancer.Lease(context);
+
+            if (next.IsError)
+            {
+                return new ErrorResponse<ServiceHostAndPort>(next.Errors);
+            }
+
+            if (!string.IsNullOrEmpty(value) && !_stored.ContainsKey(value))
+            {
+                _stored[value] = new StickySession(next.Data, DateTime.UtcNow.AddMilliseconds(_expiryInMs));
+            }
+
+            return new OkResponse<ServiceHostAndPort>(next.Data);
+        }
+
+        public void Release(ServiceHostAndPort hostAndPort)
+        {
+            
+        }
+
+        private void Expire()
+        {
+            var expired = _stored.Where(x => x.Value.Expiry < DateTime.UtcNow);
+
+            foreach(var expire in expired)
+            {
+                _stored.Remove(expire.Key, out _);
+                _loadBalancer.Release(expire.Value.HostAndPort);
+            }
+        }
+    }
+}
