+++ src/Ocelot/LoadBalancer/LoadBalancers/ILoadBalancer.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using Ocelot.Middleware;
 using Ocelot.Responses;
 using Ocelot.Values;
 
@@ -6,7 +7,7 @@ namespace Ocelot.LoadBalancer.LoadBalancers
 {
     public interface ILoadBalancer
     {
-        Task<Response<ServiceHostAndPort>> Lease();
+        Task<Response<ServiceHostAndPort>> Lease(DownstreamContext context);
         void Release(ServiceHostAndPort hostAndPort);
     }
 }
