+++ src/Ocelot/LoadBalancer/LoadBalancers/NoLoadBalancer.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Ocelot.Middleware;
 using Ocelot.Responses;
 using Ocelot.Values;
 
@@ -15,7 +16,7 @@ namespace Ocelot.LoadBalancer.LoadBalancers
             _services = services;
         }
 
-        public async Task<Response<ServiceHostAndPort>> Lease()
+        public async Task<Response<ServiceHostAndPort>> Lease(DownstreamContext downstreamContext)
         {
             //todo no point spinning a task up here, also first or default could be null..
             if (_services == null || _services.Count == 0)
