+++ test/Ocelot.AcceptanceTests/RoutingTests.cs
@@ -304,7 +304,7 @@ namespace Ocelot.AcceptanceTests
                         UpstreamPathTemplate = "/vacancy/",
                         UpstreamHttpMethod = new List<string> { "Options",  "Put", "Get", "Post", "Delete" },
                         ServiceName = "botCore",
-                        LoadBalancer = "LeastConnection"
+                        LoadBalancerOptions = new FileLoadBalancerOptions { Type = "LeastConnection" }
                     },
                     new FileReRoute
                     {
@@ -321,7 +321,7 @@ namespace Ocelot.AcceptanceTests
                         UpstreamPathTemplate = "/vacancy/{vacancyId}",
                         UpstreamHttpMethod = new List<string> { "Options",  "Put", "Get", "Post", "Delete" },
                         ServiceName = "botCore",
-                        LoadBalancer = "LeastConnection"
+                        LoadBalancerOptions = new FileLoadBalancerOptions { Type = "LeastConnection" }
                     }
                 }
             };
@@ -802,7 +802,7 @@ namespace Ocelot.AcceptanceTests
                         UpstreamPathTemplate = "/vacancy/",
                         UpstreamHttpMethod = new List<string> { "Options",  "Put", "Get", "Post", "Delete" },
                         ServiceName = "botCore",
-                        LoadBalancer = "LeastConnection"
+                        LoadBalancerOptions = new FileLoadBalancerOptions { Type = "LeastConnection" }
                     },
                     new FileReRoute
                     {
@@ -819,7 +819,7 @@ namespace Ocelot.AcceptanceTests
                         UpstreamPathTemplate = "/vacancy/{vacancyId}",
                         UpstreamHttpMethod = new List<string> { "Options",  "Put", "Get", "Post", "Delete" },
                         ServiceName = "botCore",
-                        LoadBalancer = "LeastConnection"
+                        LoadBalancerOptions = new FileLoadBalancerOptions { Type = "LeastConnection" }
                     }
                 }
             };
