+++ test/Ocelot.UnitTests/LoadBalancer/LeastConnectionTests.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Ocelot.LoadBalancer.LoadBalancers;
+using Ocelot.Middleware;
 using Ocelot.Responses;
 using Ocelot.Values;
 using Shouldly;
@@ -17,9 +19,11 @@ namespace Ocelot.UnitTests.LoadBalancer
         private LeastConnection _leastConnection;
         private List<Service> _services;
         private Random _random;
+        private DownstreamContext _context;
 
         public LeastConnectionTests()
         {
+            _context = new DownstreamContext(new DefaultHttpContext());
             _random = new Random();
         }
 
@@ -60,9 +64,9 @@ namespace Ocelot.UnitTests.LoadBalancer
 
             _leastConnection = new LeastConnection(() => Task.FromResult(availableServices), serviceName);
 
-            var hostAndPortOne = _leastConnection.Lease().Result;
+            var hostAndPortOne = _leastConnection.Lease(_context).Result;
             hostAndPortOne.Data.DownstreamHost.ShouldBe("127.0.0.1");
-            var hostAndPortTwo = _leastConnection.Lease().Result;
+            var hostAndPortTwo = _leastConnection.Lease(_context).Result;
             hostAndPortTwo.Data.DownstreamHost.ShouldBe("127.0.0.2");
             _leastConnection.Release(hostAndPortOne.Data);
             _leastConnection.Release(hostAndPortTwo.Data);
@@ -72,9 +76,9 @@ namespace Ocelot.UnitTests.LoadBalancer
                 new Service(serviceName, new ServiceHostAndPort("127.0.0.1", 80), string.Empty, string.Empty, new string[0]),
             };
 
-            hostAndPortOne = _leastConnection.Lease().Result;
+            hostAndPortOne = _leastConnection.Lease(_context).Result;
             hostAndPortOne.Data.DownstreamHost.ShouldBe("127.0.0.1");
-            hostAndPortTwo = _leastConnection.Lease().Result;
+            hostAndPortTwo = _leastConnection.Lease(_context).Result;
             hostAndPortTwo.Data.DownstreamHost.ShouldBe("127.0.0.1");
             _leastConnection.Release(hostAndPortOne.Data);
             _leastConnection.Release(hostAndPortTwo.Data);
@@ -85,9 +89,9 @@ namespace Ocelot.UnitTests.LoadBalancer
                 new Service(serviceName, new ServiceHostAndPort("127.0.0.2", 80), string.Empty, string.Empty, new string[0]),
             };
 
-            hostAndPortOne = _leastConnection.Lease().Result;
+            hostAndPortOne = _leastConnection.Lease(_context).Result;
             hostAndPortOne.Data.DownstreamHost.ShouldBe("127.0.0.1");
-            hostAndPortTwo = _leastConnection.Lease().Result;
+            hostAndPortTwo = _leastConnection.Lease(_context).Result;
             hostAndPortTwo.Data.DownstreamHost.ShouldBe("127.0.0.2");
             _leastConnection.Release(hostAndPortOne.Data);
             _leastConnection.Release(hostAndPortTwo.Data);
@@ -95,7 +99,7 @@ namespace Ocelot.UnitTests.LoadBalancer
 
         private async Task LeaseDelayAndRelease()
         {
-            var hostAndPort = await _leastConnection.Lease();
+            var hostAndPort = await _leastConnection.Lease(_context);
             await Task.Delay(_random.Next(1, 100));
             _leastConnection.Release(hostAndPort.Data);
         }
@@ -134,15 +138,15 @@ namespace Ocelot.UnitTests.LoadBalancer
             _services = availableServices;
             _leastConnection = new LeastConnection(() => Task.FromResult(_services), serviceName);
 
-            var response = _leastConnection.Lease().Result;
+            var response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[0].HostAndPort.DownstreamHost);
 
-            response = _leastConnection.Lease().Result;
+            response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[1].HostAndPort.DownstreamHost);
 
-            response = _leastConnection.Lease().Result;
+            response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[2].HostAndPort.DownstreamHost);
         }
@@ -161,19 +165,19 @@ namespace Ocelot.UnitTests.LoadBalancer
             _services = availableServices;
             _leastConnection = new LeastConnection(() => Task.FromResult(_services), serviceName);
 
-            var response = _leastConnection.Lease().Result;
+            var response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[0].HostAndPort.DownstreamHost);
 
-            response = _leastConnection.Lease().Result;
+            response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[1].HostAndPort.DownstreamHost);
 
-            response = _leastConnection.Lease().Result;
+            response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[0].HostAndPort.DownstreamHost);
 
-            response = _leastConnection.Lease().Result;
+            response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[1].HostAndPort.DownstreamHost);
         }
@@ -192,26 +196,26 @@ namespace Ocelot.UnitTests.LoadBalancer
             _services = availableServices;
             _leastConnection = new LeastConnection(() => Task.FromResult(_services), serviceName);
 
-            var response = _leastConnection.Lease().Result;
+            var response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[0].HostAndPort.DownstreamHost);
 
-            response = _leastConnection.Lease().Result;
+            response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[1].HostAndPort.DownstreamHost);
 
-            response = _leastConnection.Lease().Result;
+            response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[0].HostAndPort.DownstreamHost);
 
-            response = _leastConnection.Lease().Result;
+            response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[1].HostAndPort.DownstreamHost);
 
             //release this so 2 should have 1 connection and we should get 2 back as our next host and port
             _leastConnection.Release(availableServices[1].HostAndPort);
 
-            response = _leastConnection.Lease().Result;
+            response = _leastConnection.Lease(_context).Result;
 
             response.Data.DownstreamHost.ShouldBe(availableServices[1].HostAndPort.DownstreamHost);
         }
@@ -272,7 +276,7 @@ namespace Ocelot.UnitTests.LoadBalancer
 
         private void WhenIGetTheNextHostAndPort()
         {
-            _result = _leastConnection.Lease().Result;
+            _result = _leastConnection.Lease(_context).Result;
         }
 
         private void ThenTheNextHostAndPortIsReturned()
